refactor(FontSettingsSelectionDelegate): changed method parameter from 'CGFloat' to 'Float'.

+++ NovelReader/Screens/OptionsView/Protocol/FontSettingsSelectionDelegate.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 protocol FontSettingsSelectionDelegate: AnyObject {
-    func didSetFontSize(to newSize: CGFloat)
+    func didSetFontSize(to newSize: Float)
     func didSetFontFamily(to fontFamily: String)
 }
